requirements.txt
nose
sphinx
python-telegram-bot
requests
